requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
openpyxl==3.1.2
PyJWT==2.8.0
bcrypt==4.1.2
redis==5.0.1